FIX: Two PR-related fixes
- two bug-fixes related to PR #186

 On branch conn_doc_tests
 Changes to be committed:
	modified:   syncopy/connectivity/connectivity_analysis.py
	modified:   syncopy/tests/synth_data.py

+++ syncopy/tests/synth_data.py
@@ -139,7 +139,7 @@ def mk_RandomAdjMat(nChannels=3, conn_thresh=0.25, max_coupling=0.25):
     AdjMat = np.random.random_sample((nChannels, nChannels))
 
     # all smaller than threshold elements get set to 1 (coupled)
-    AdjMat = (AdjMat < conn_thresh)
+    AdjMat = (AdjMat < conn_thresh).astype(float)
 
     # set diagonal to 0 to easier identify coupling
     np.fill_diagonal(AdjMat, 0)
